Fixed nondeterministic trace file output bug
Segments were being written (for debugging and validation, eg., comparing exchange and
iexchange outputs) incorrectly. In ptrace.cpp,
write_segments was being called on block(0), but this is local block 0
of the current MPI rank. The needs to be global (gid) block 0. Enclosing
this call in a test for world.rank() == 0 fixes this problem (only
writing from local block 0 of rank 0).