Added barrier to start of particle tracing
Removes skew from generating or reading data prior to starting particle
tracing.
Having the particle tracing start out synchronized makes performance
easier to profile, visualize, and understand.
The barrier is profiled so that it appears in the profvis trace, as a
visual to see how closely the clocks are synchronized.